Fix mpi logic in CABLE SPD
Currently the CABLE spack build leaves the `CABLE_MPI` CMake variable
unset when `~mpi` is satisfied and relies on the default value for
`CABLE_MPI`. This change removes the if condition around
`define_from_variant` so that `CABLE_MPI` is specified to CMake
regardless of how the mpi variant is set.

Fixes #76