Spack Packages for Each Component of ACCESS-OM3 (#171)
Introduce seperate spack packages for the components of ACCESS-OM3. This adds packages for MOM6, CICE, WW3 and access3-share. There is also an an access3 package, which shares a source repository with access3-share, to build the single model executable for OM3. The access3 package can be extended to include the UM in the future. These packages replace the access-om3-nuopc package, which can be deprecated in the future. The packages for individual models can be used for "standalone" model builds too.

---------

Co-authored-by: minghangli-uni <minghang.li1@anu.edu.au>
Co-authored-by: Dougie Squire <42455466+dougiesquire@users.noreply.github.com>